feat: Add support for open command
Added three new keys, 'open = ' is the same as 'editor = ', and 'o = '
tells trans what command to run when key o is pressed, with the cursor on a
file; 'enter = ' is for enter, similar to o. 'o = ' and 'enter = ' will
override 'editor = ' and 'open = ', on o and enter keys repectively.

When 'editor = ' and 'open = ' are both present, the last one will be
selected.

Also supports command line arguments, for 'editor = ' as well.

Signed-off-by: Howard Chu <howardchu95@gmail.com>