general: Add more tests, and secure getting terminal height using get_height()
Add read_content(), bottom(), up(), down(), left(), and right() tests.
Originally I used height - 1 as actually height, but when testing,
height can be 0, so height - 1 causes an unsigned integer underflow; To
avoid this, I refactored all the height retrievals into a helper function
get_height(), which checks if the height is 0 before subtracting, this
reduces the risk of underflow.

I also changed the name of the variable that contains all the contents in
the current directory to 'content', it used to be 'current_dir', which
is misleading, because it does not represent a directory of any kind.

Signed-off-by: Howard Chu <howardchu95@gmail.com>