canvas: Reduce thousands of syscalls to two-digit
When run with strace, there are hundreds of thousands of system calls
invoked by the draw() function. This happens because is_dir() is called
for every pixel on the canvas. This is utterly stupid and severely
impacts performance. Add checks before calling is_dir() to reduce the
number of system calls.

Signed-off-by: Howard Chu <howardchu95@gmail.com>